Cloud Function-SD-PubSub/Q3/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.25.0